full refactor of fundraising and fundraising purchase  :sparkles:  :boom:  :bricks:  :recycle:

+++ backend/src/modules/fundraising/use-cases/fundraising/delete-fundraising/delete-fundraising-usecase.ts
@@ -4,6 +4,12 @@ class DeleteFundraisingUseCase {
   constructor(private fundraisingRepository: IFundraisingRepository) {}
 
   public async execute(id: string): Promise<void> {
+    const fundraising = await this.fundraisingRepository.findById(id)
+
+    if (!fundraising) {
+      throw new Error('Fundraising not found')
+    }
+
     await this.fundraisingRepository.delete(id)
   }
 }
